requirements.txt
attrs==22.2.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
greenlet==2.0.1
idna==3.4
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql==0.0.3
mysql-connector-python==8.0.29
mysqlclient==2.1.1
packaging==22.0
pluggy==1.0.0
protobuf==3.20.1
psycopg2==2.9.5
PyJWT==2.6.0
pytest==7.2.0
python-dotenv==0.21.0
requests==2.28.1
SQLAlchemy==1.4.46
urllib3==1.26.13
Werkzeug==2.2.2
